refactor: change chatreactions to work like guild shit
